fix(transport): relay reputation probationary window and bridge adjustment types
- AndroidPlatformBridge.kt: use UInt? and Byte? for bleAdjustment caches matching UniFFI generated types

- mesh_routing.rs: grant 3-attempt probationary window before blackballing relay during transient handoffs

- test: add unit test test_reputation_probationary_period (passed)

- governance: Rule 8 adversarial approval bod-4df59504 recorded in HANDOFF/review/

+++ android/app/src/main/java/com/scmessenger/android/service/AndroidPlatformBridge.kt
@@ -82,9 +82,9 @@ class AndroidPlatformBridge @Inject constructor(
     @Volatile private var hasWifi: Boolean = false
     @Volatile private var hasCellular: Boolean = false
     @Volatile private var currentMotionState: uniffi.api.MotionState = uniffi.api.MotionState.UNKNOWN
-    @Volatile private var lastScanIntervalMs: Long? = null
-    @Volatile private var lastAdvertiseIntervalMs: Long? = null
-    @Volatile private var lastTxPowerDbm: Int? = null
+    @Volatile private var lastScanIntervalMs: UInt? = null
+    @Volatile private var lastAdvertiseIntervalMs: UInt? = null
+    @Volatile private var lastTxPowerDbm: Byte? = null
     @Volatile private var lastRelayMaxPerHour: UInt? = null
 
     /**
